Indicate the current CheriBSD version.
Not only the third-party packages but also other sections (e.g.,
CheriBSD features) might depend on a specific CheriBSD version.

Places where the CheriBSD version is noted are annotated with comments
reminding other places where the version is mentioned.
We could use an mdBook preprocessor to define a variable with a version
and use it throughout the guide. However, README.md rendered by GitHub
(not GitHub Pages) would include "{{ version }}" instead of a variable.

Reviewed by:	rwatson