Use FAT32 for an SD card when posible
Despite of the recommendations from Arm [0], FAT32 works fine for an SD
card storing a firmware. Use FAT32 instead of FAT16 to partition an SD
card larger than the maximum allowed partition size in FAT16 (2GB or
4GB).

On macOS, diskutil can determine the right partition type for an SD card
depending on its size. However, on Linux, parted doesn't allow to
specify "fat as a filesystem but requires "fat16" or "fat32".

[0] https://developer.arm.com/documentation/den0132/0100/Flash-the-onboard-SD-card